Add 'koa-' prefix to middleware module name
