+++ packages/lit-ssr/src/lib/render-html-file.ts
@@ -1,73 +1,84 @@
-import Koa from 'koa';
-import {renderModule} from './render-module.js';
-import {IterableReader} from './util/iterator-readable.js';
+import {render} from './render-lit-html.js';
+import {$private} from 'lit-html/private-ssr-support.js';
+import {promises as fs} from 'fs';
+import path from 'path';
+import fetch from 'node-fetch';
+import {URL} from 'url';
 
-import * as path from 'path';
+const {HTML_RESULT} = $private;
 
-export interface RenderHTMLFileOptions {
-  /** The HTTP root folder, for resolving URL paths to static files */
+interface RenderAppOptions {
+  url: string;
   root: string;
-  /** A fallback file to use if the specified URL did not exist */
   fallback: string;
+  env: {[key: string]: unknown};
+}
+
+interface InitializeSSRModule {
+  initializeSSR?: (location: string) => void | Promise<unknown>;
 }
 
 /**
- * Koa middleware for rendering HTML files using render-lit-html.
- *
- * This middleware will only handle URLs to files with `.html` or bare paths
- * with `accept: text/html` or `accept: *`. If a static file does not exist at
- * the specified URL path, the `fallback` file will be rendered/served.
- *
- * Any `<script>` tags that should be loaded & run on the server prior to
- * rendering should be marked with an `ssr` attribute.  If any of these scripts
- * are modules that have an `initializeSSR` export, it will be invoked and
- * awaited prior to rendering.  If any `initializeSSR` methods resolve to arrays
- * of values, those values will be interpolated into `<!--lit-ssr-value-->`
- * comment markers found in the page.
- *
- * @param options
+ * Process to render an HTML file for a given URL:
+ * 1. Read the specified HTML file
+ * 2. Find all `<script>` tags marked with `ssr` attribute and dynamically
+ *    import them
+ * 3. If the module exported an `initializeSSR` method, call and await its
+ *    result
+ * 4. If `initializeSSR` returned a values array, push those values onto the the
+ *    dynamic values to render the page with
+ * 5. Split the HTML file contents on the `<!--lit-ssr-value-->` markers
+ * 6. Return a TemplateResult constructed from the HTML page's contents and the
+ *    values returned from initializeSSR.
  */
-export const renderHTMLFile = (options: RenderHTMLFileOptions) => async (
-  ctx: Koa.ParameterizedContext<Koa.DefaultState, Koa.DefaultContext>,
-  next: Koa.Next
-) => {
-  // Only handle GET's
-  if (ctx.method !== 'GET') {
-    return next();
-  }
-  // Only handle .html extension or bare paths
-  const ext = path.extname(ctx.path);
-  if (ext && ext !== '.html') {
-    return next();
-  }
-  // Only handle HTML or *, but never JSON
-  if (!ctx.headers || typeof ctx.headers.accept !== 'string') {
-    return next();
-  }
-  if (ctx.headers.accept.includes('application/json')) {
-    return next();
-  }
-  if (
-    !(
-      ctx.headers.accept.includes('text/html') ||
-      ctx.headers.accept.includes('*/*')
-    )
-  ) {
-    return next();
-  }
-  // Render the file using a new VM context for each request
-  const ssrResult = await renderModule(
-    './render-html-file-impl.js',
-    import.meta.url,
-    'renderFile',
-    [{url: ctx.href, root: options.root, fallback: options.fallback}]
-  );
-  const stream = new IterableReader(ssrResult);
-  stream.on('error', (e) => {
-    console.error(e.message);
-    console.error(e.stack);
+export async function renderFile(options: RenderAppOptions) {
+  // Shim some useful globals onto window
+  const url = new URL(options.url);
+  Object.assign(window, {
+    location: url,
+    fetch,
+    process: {env: {NODE_ENV: 'production', ...(options.env || {})}},
   });
-  ctx.type = 'text/html';
-  ctx.body = stream;
-  return;
-};
+  // Make sure file exists; if not, use fallback
+  let file = path.join(options.root, url.pathname);
+  let exists = false;
+  try {
+    exists = (await fs.stat(file)).isFile();
+  } catch {
+    /* Use fallback */
+  }
+  if (!exists) {
+    file = path.join(options.root, options.fallback);
+  }
+  // Read file
+  const content = await fs.readFile(file, 'utf-8');
+  // Load `ssr`-tagged scripts
+  const scripts = Array.from(
+    content.matchAll(/<script[^>]* src="([^"]+)" ssr>/g)
+  ).map((m) => m[1]);
+  const values = [];
+  for (const script of scripts) {
+    const module = await import(path.join(options.root, script));
+    if ((module as InitializeSSRModule).initializeSSR) {
+      const ret = await module.initializeSSR(url);
+      if (Array.isArray(ret)) {
+        values.push(...ret);
+      }
+    }
+  }
+  // Split strings for any interpolated values
+  const strings = content.split('<!--lit-ssr-value-->');
+  if (strings.length - 1 !== values.length) {
+    throw new Error(
+      `Number of <!--lit-ssr-value--> comments (${
+        strings.length - 1
+      }) and initializeSSR()-returned values (${values.length}) did not match.`
+    );
+  }
+  // Construct a TemplateResult
+  return render({
+    _$litType$: HTML_RESULT,
+    strings,
+    values,
+  });
+}
